requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
